feat: move Card component to CardsList dir
