+++ src/components/CardsList/Card/Card.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { CharacterSpecies, CharacterStatus } from '../../types/enums';
+import { CharacterSpecies, CharacterStatus } from '../../../types/enums';
 import { IMainCard } from 'types/interfaces';
 import styles from './Card.module.css';
-import aliveImage from '../../assets/images/alive.png';
-import deadImage from '../../assets/images/dead.png';
-import alienImage from '../../assets/images/alien.png';
-import humanImage from '../../assets/images/human.png';
+import aliveImage from '../../../assets/images/alive.png';
+import deadImage from '../../../assets/images/dead.png';
+import alienImage from '../../../assets/images/alien.png';
+import humanImage from '../../../assets/images/human.png';
 
 interface ICardProps {
   card: IMainCard;
